feat: Adds plugin manager
