+++ main.go
@@ -1,6 +1,8 @@
 package main
 
-import "github.com/kaytu-io/kaytu/cmd"
+import (
+	"github.com/kaytu-io/kaytu/cmd"
+)
 
 func main() {
 	cmd.Execute()
