+++ pkg/api/github/entity.go
@@ -1,0 +1,38 @@
+package github
+
+import "time"
+
+type Release struct {
+	Url             string    `json:"url"`
+	AssetsUrl       string    `json:"assets_url"`
+	UploadUrl       string    `json:"upload_url"`
+	HtmlUrl         string    `json:"html_url"`
+	Id              int       `json:"id"`
+	NodeId          string    `json:"node_id"`
+	TagName         string    `json:"tag_name"`
+	TargetCommitish string    `json:"target_commitish"`
+	Name            string    `json:"name"`
+	Draft           bool      `json:"draft"`
+	Prerelease      bool      `json:"prerelease"`
+	CreatedAt       time.Time `json:"created_at"`
+	PublishedAt     time.Time `json:"published_at"`
+	Assets          []Asset   `json:"assets"`
+	TarballUrl      string    `json:"tarball_url"`
+	ZipballUrl      string    `json:"zipball_url"`
+	Body            string    `json:"body"`
+}
+
+type Asset struct {
+	Url                string    `json:"url"`
+	Id                 int       `json:"id"`
+	NodeId             string    `json:"node_id"`
+	Name               string    `json:"name"`
+	Label              string    `json:"label"`
+	ContentType        string    `json:"content_type"`
+	State              string    `json:"state"`
+	Size               int       `json:"size"`
+	DownloadCount      int       `json:"download_count"`
+	CreatedAt          time.Time `json:"created_at"`
+	UpdatedAt          time.Time `json:"updated_at"`
+	BrowserDownloadUrl string    `json:"browser_download_url"`
+}
