--- pkg/aws/config.go
@@ -1,51 +1,0 @@
-package aws
-
-import (
-	"fmt"
-	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/config"
-	"github.com/aws/aws-sdk-go-v2/credentials"
-	"github.com/aws/aws-sdk-go-v2/credentials/stscreds"
-	"github.com/aws/aws-sdk-go-v2/service/sts"
-	"golang.org/x/net/context"
-)
-
-func GetConfig(ctx context.Context, awsAccessKey, awsSecretKey, awsSessionToken, assumeRoleArn string, profile *string, externalId *string) (aws.Config, error) {
-	opts := []func(*config.LoadOptions) error{}
-
-	if awsAccessKey != "" {
-		opts = append(opts, config.WithCredentialsProvider(credentials.NewStaticCredentialsProvider(awsAccessKey, awsSecretKey, awsSessionToken)))
-	}
-
-	if profile != nil && *profile != "" {
-		opts = append(opts, config.WithSharedConfigProfile(*profile))
-	}
-
-	cfg, err := config.LoadDefaultConfig(ctx, opts...)
-	if err != nil {
-		return aws.Config{}, fmt.Errorf("failed to load AWS config: %w", err)
-	}
-	if cfg.Region == "" {
-		cfg.Region = "us-east-1"
-	}
-
-	if assumeRoleArn != "" {
-		cfg, err = config.LoadDefaultConfig(
-			context.Background(),
-			config.WithCredentialsProvider(
-				stscreds.NewAssumeRoleProvider(
-					sts.NewFromConfig(cfg),
-					assumeRoleArn,
-					func(o *stscreds.AssumeRoleOptions) {
-						o.ExternalID = externalId
-					},
-				),
-			),
-		)
-		if err != nil {
-			return aws.Config{}, fmt.Errorf("failed to assume role: %w", err)
-		}
-	}
-
-	return cfg, nil
-}
