+++ pkg/server/config.go
@@ -4,45 +4,85 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/golang-jwt/jwt/v4"
+	"github.com/kaytu-io/kaytu/pkg/plugin/proto/src/golang"
 	"os"
 	"strings"
+	"sync"
 	"time"
 )
 
+type Plugin struct {
+	Config *golang.RegisterConfig `json:"config"`
+}
+
+func (p *Plugin) Path() string {
+	return PluginDir() + p.Config.Name
+}
+
 type Config struct {
-	AccessToken      string `json:"access-token"`
-	DefaultWorkspace string `json:"default_workspace"`
+	AccessToken string    `json:"access-token"`
+	Plugins     []*Plugin `json:"plugins"`
 }
 
 var ExpiredSession = fmt.Errorf("your session has expired, please login again using `kaytu login`")
 
-func GetConfig() (*Config, error) {
+var (
+	once      sync.Once
+	config    *Config
+	configErr error
+)
+
+func GetPlugins() ([]*Plugin, error) {
+	var cfg Config
 	home := os.Getenv("HOME")
 	data, err := os.ReadFile(home + "/.kaytu/kaytu-config.json")
 	if err != nil {
 		if strings.Contains(err.Error(), "no such file or directory") {
 			return nil, nil
 		}
-		return nil, fmt.Errorf("[CredentialsFile] : %v", err)
+		return nil, fmt.Errorf("[GetPlugins] : %v", err)
 	}
 
+	err = json.Unmarshal(data, &cfg)
+	if err != nil {
+		return nil, fmt.Errorf("[GetPlugins] : %v", err)
+	}
+
+	return cfg.Plugins, nil
+}
+
+func GetConfig() (*Config, error) {
+	once.Do(func() {
+		config, configErr = loadConfig()
+	})
+	return config, configErr
+}
+
+func loadConfig() (*Config, error) {
 	var config Config
+	home := os.Getenv("HOME")
+	data, err := os.ReadFile(home + "/.kaytu/kaytu-config.json")
+	if err != nil {
+		if strings.Contains(err.Error(), "no such file or directory") {
+			return &config, nil
+		}
+		return nil, fmt.Errorf("[loadConfig] : %v", err)
+	}
+
 	err = json.Unmarshal(data, &config)
 	if err != nil {
-		return nil, fmt.Errorf("[getConfig] : %v", err)
+		return nil, fmt.Errorf("[loadConfig] : %v", err)
 	}
 
-	if config.AccessToken == "" {
-		return nil, fmt.Errorf("please log in first")
-	}
+	if config.AccessToken != "" {
+		checkEXP, err := CheckExpirationTime(config.AccessToken)
+		if err != nil {
+			return nil, fmt.Errorf("[loadConfig] : %v", err)
+		}
 
-	checkEXP, err := CheckExpirationTime(config.AccessToken)
-	if err != nil {
-		return nil, fmt.Errorf("[getConfig] : %v", err)
-	}
-
-	if checkEXP == true {
-		return nil, ExpiredSession
+		if checkEXP == true {
+			config.AccessToken = ""
+		}
 	}
 	return &config, nil
 }
