+++ pkg/utils/hash.go
@@ -1,0 +1,11 @@
+package utils
+
+import (
+	"crypto/sha256"
+	"encoding/hex"
+)
+
+func HashString(str string) string {
+	h := sha256.New()
+	return hex.EncodeToString(h.Sum([]byte(str)))
+}
